bootcamp ownership, course ownership, protecting unauthorized actions on bootcamps and courses

+++ controllers/bootcamps.js
@@ -28,7 +28,25 @@ const getBootcamp = asyncHandler(async (req, res, next) => {
 // route: POST /api/v1/bootcamps/:id
 // method: Private
 const createBootcamp = asyncHandler(async (req, res, next) => {
+  // add user to req.body
+  req.body.user = req.user
+
+  // check for published bootcamp
+  const publishedBootcamp = await Bootcamp.findOne({ user: req.user.id })
+
+  // if the user is not an admin they can only add one bootcamp
+
+  if (publishedBootcamp && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `The user with id ${req.user.id} has already published a bootcamp`,
+        400
+      )
+    )
+  }
+
   const bootcamp = await Bootcamp.create(req.body)
+
   res.status(201).json({
     success: true,
     data: bootcamp,
@@ -39,15 +57,25 @@ const createBootcamp = asyncHandler(async (req, res, next) => {
 // route: PUT /api/v1/bootcamps/:id
 // method: Private
 const updateBootcamp = asyncHandler(async (req, res, next) => {
-  const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  })
+  let bootcamp = await Bootcamp.findById(req.params.id)
   if (!bootcamp) {
     return next(
       new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
     )
   }
+  // make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.name} is not authorized to update this bootcamp`,
+        401
+      )
+    )
+  }
+  bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true,
+  })
   res.status(200).json({ success: true, data: bootcamp })
 })
 
@@ -59,6 +87,15 @@ const deleteBootcamp = asyncHandler(async (req, res, next) => {
   if (!bootcamp) {
     return next(
       new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+    )
+  }
+  // make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.name} is not authorized to delete this bootcamp`,
+        401
+      )
     )
   }
   bootcamp.deleteOne()
@@ -99,6 +136,15 @@ const bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
   if (!bootcamp) {
     return next(
       new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+    )
+  }
+  // make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.name} is not authorized to update this bootcamp`,
+        401
+      )
     )
   }
   if (!req.files) {
